Improved the layout so that only the content is scrolled when it is to big to be displayed on the page
