+++ src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/[configId]/page.tsx
@@ -14,7 +14,7 @@ const MachineConfigView: React.FC<MachineConfigProps> = async ({ params: { confi
   //replace ConfigContent <-> MachineConfigEditor as needed
   return (
     <Content title={`Tech Data Set: ${machineConfig.name}`}>
-      <ConfigPage originalParentConfig={machineConfig} />
+      <ConfigPage parentConfig={machineConfig} />
     </Content>
   );
 };
